refactor: extract styles from Container to separate file
