+++ src/App.jsx
@@ -1,5 +1,5 @@
 import Header from "./components/Header/Header";
-import Container from "./components/Container";
+import Container from "./components/Container/Container";
 import Card from "./components/Card/Card";
 
 const App = () => {
