+++ src/components/Container/Container.jsx
@@ -1,0 +1,7 @@
+import { StyledContainer } from "./ContainerStyles";
+
+const Container = ({ children }) => {
+  return <StyledContainer>{children}</StyledContainer>;
+};
+
+export default Container;
