+++ src/components/Container/ContainerStyles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledContainer = styled.main`
+  max-width: 600px;
+  margin: 2rem auto;
+  padding: 1rem;
+`;
